Tests passing, directory restructure
