+++ test/test.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <math.h>
 #include <assert.h>
-#include "../fft.h"
+#include "fft.h"
 
 #define TEST_1_SAMPLE_RATE      (500)
 #define TEST_2_SAMPLE_RATE      (2000)
@@ -12,7 +12,7 @@
 
 #define TEST_1_FREQ             (20)
 #define TEST_2_FREQ_1           (50)
-#define TEST_2_FREQ_2           (500)
+#define TEST_2_FREQ_2           (100)
 
 #define PASS                    (0)
 #define FAIL                    (1)
@@ -57,7 +57,7 @@ void print_samples(dig_t *samples, int num_samples)
 void print_fft_output(sample_t *output, int num_samples) 
 {
     printf("Printing FFT output:\n");
-    for (int i = 0; i < num_samples; i++) 
+    for (int i = 0; i < num_samples/2; i++) 
     {
         printf("%d,%f, %f\n", i, creal(output[i]), cimag(output[i]));
     }
@@ -83,11 +83,10 @@ void test_fft()
     fft();
     freq_t detected_freq2 = find_freq(output_buffer, TEST_2_SAMPLE_RATE);
     printf("Expected primary freq for two sine waves: %d, Detected: %f\n", TEST_2_FREQ_1, detected_freq2);
-    assert(fabs(detected_freq - TEST_1_FREQ) <= FREQ_TOLERANCE(TEST_2_SAMPLE_RATE));
-    int f2_bucket = (FFT_BUFFER_SIZE * TEST_2_FREQ_2) / TEST_2_SAMPLE_RATE;
-    double magnitude_diff = fabs((double)output_buffer[f2_bucket] - (double)FREQ_2_MAG);
-    printf("Expected magnitude for secondary freq: %d, Detected magnitude difference: %f\n", FREQ_2_MAG, magnitude_diff);
-    assert(magnitude_diff <= MAG_TOLERANCE);
+    assert(fabs(detected_freq2 - TEST_2_FREQ_1) <= FREQ_TOLERANCE(TEST_2_SAMPLE_RATE));
+    int f2_bucket = (int)round(((double)FFT_BUFFER_SIZE * (double)TEST_2_FREQ_2) / (double)TEST_2_SAMPLE_RATE);
+    printf("Magnitude for secondary freq %d (bucket %d): %f\n", TEST_2_FREQ_2, f2_bucket, (float)output_buffer[f2_bucket]);
+    assert((double)output_buffer[f2_bucket] >= FREQ_2_MAG * 1000); // FFT output not to scale
 }
